requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob==12.9.0
tensorflow==2.1.0
tf-bodypix
tfjs-graph-converter
numpy
Pillow
opencv-python
opencv-contrib-python
